+++ lib/src/utils.dart
@@ -74,6 +74,36 @@ bool debugCheckHasFluentLocalizations(BuildContext context) {
   return true;
 }
 
+/// Asserts that the given context has a [Flyout] ancestor.
+///
+/// To call this function, use the following pattern, typically in the
+/// relevant Widget's build method:
+///
+/// ```dart
+/// assert(debugCheckHasFlyout(context));
+/// ```
+///
+/// Does nothing if asserts are disabled. Always returns true.
+bool debugCheckHasFlyout(BuildContext context) {
+  assert(() {
+    if (Flyout.maybeOf(context) == null) {
+      throw FlutterError.fromParts(<DiagnosticsNode>[
+        ErrorSummary('No Flyout found.'),
+        ErrorDescription(
+          '${context.widget.runtimeType} widgets require a Flyout '
+          'to be provided by a Flyout widget ancestor.',
+        ),
+        ErrorHint(
+          'To introduce a Flyout, use the showFlyout method or related methods.',
+        ),
+        ...context.describeMissingAncestor(expectedAncestorType: Flyout)
+      ]);
+    }
+    return true;
+  }());
+  return true;
+}
+
 /// Check if the current screen is 10 foot long or bigger.
 ///
 /// [width] is the width of the current screen. If not provided,